feat: supporting bucket specific secrets for aws and google
To use, add a secret prefixed with the name of the bucket and a hyphen.
E.g. if your bucket is mybucket, your secret name would be
`mybucket-google-secret.json` or `mybucket-aws-secret.json`.

The logic will default to `google-secret.json` or `aws-secret.json`
if there is no specific bucket secret.

NEW: If no secret is provided at all, use default application credentials
for Google.

Resolves #105 and partially #35

It's easy to implement for AWS, but I'm not set up to test it.