perf: 1x memory usage for parallel downloads (#112)
Previously, I was very conservative and wanted to ensure
that the usage of shared memory was completely invisible
to the end user of downloaded numpy arrays. However, this
required making a copy resulting in 2x memory usage.

I conducted some experiments showing that we can safely
avoid making this copy. Instead, we merely unlink the
shared memory. The memory region will remain accessible
so long as there is an open file descriptor to it. The FD
will be closed when the output VolumeCutout is garbage collected.

In all common use cases I have witnessed, this should result in
the transparent usage of shared memory without the concomitant
memory cost. In the case that a user fills an array with
thousands of VolumeCutouts, they have the option of making a
copy using np.copy and closing the file descriptors manually
or automatically.