redesign: switch to using .shard from .index + .data
Per personal communication dated Aug. 29, JBMS indicated
that .shard files are in and .index and .data files are out
due to the difficulty of atomic updates. The shard index is now
concatenated at the front and the offsets in the shard index and
minishard index are now implicitly offset by the size of the shard
index.