perf: instant fs based shared memory files + use spare file capability
Previously we were writing out real files full of zeros,
but a) that takes forever and b) modern OSes support sparse
files (i.e. gigabytes of zeros that take up only a few real bytes).

The following should take advantage of this functionality and be faster.
It should work on Python 2.7 Unix, and Python 3.5+ on Unix and Windows.

References:
  https://stackoverflow.com/questions/8816059/create-file-of-particular-size-in-python
  https://docs.python.org/3/library/os.html#os.ftruncate
  https://docs.python.org/2/library/os.html#os.ftruncate