Do not delay creation of metadata when using dask
I originally made this change before I understood the dask ecosystem, and I was desiring to have many volumes'  created in parallel across many workers.

For other output formats (parquet, zarr, etc), the metadata is written at graph creation time, not delayed to run time.  So this change modifies the behavior to be consistent with the rest of the ecosystem.