Revert delay of creation of metadata when using dask (#498)
* Fix and improve testing
- Appropriate check_meta/check_type (fixes some breakages due to different underlying chunk types)
- Use random integers for arrays instead of all zeros

* Do not delay creation of metadata when using dask

I originally made this change before I understood the dask ecosystem, and I was desiring to have many volumes'  created in parallel across many workers.

For other output formats (parquet, zarr, etc), the metadata is written at graph creation time, not delayed to run time.  So this change modifies the behavior to be consistent with the rest of the ecosystem.